refactored and simplified DataStore in Problem, ProblemDataStore
Former-commit-id: 9c27ba5dac67fc1bf05375d5f4f42a7e40573455 [formerly 7a0bbcf31b897c909556d1d532b41f059e5b3a5c [formerly e12f6d9c89d4a8d693a3f7f5d8b4ea988adc5245]]
Former-commit-id: 2ba3b49ae13e4396edadb286fa2938267bc9448b
Former-commit-id: 825895bf5aece696a979754beaf6806fd80f2071

+++ artap/datastore.py
@@ -108,23 +108,28 @@ class DataStore:
 
 
 class SqliteDataStore(DataStore):
-    def __init__(self, problem=None, database_file=None, create_database=False):
+    def __init__(self, problem, database_name=None, remove_existing=True):
         super().__init__(problem)
 
-        if self.problem is not None:
-            problem.data_store = self
+        self.database_name = database_name
 
-        if create_database:
-            self.database_name = self.problem.working_dir + os.sep + "data" + ".sqlite"
-            if os.path.exists(self.database_name):
-                os.remove(self.database_name)
-        else:
-            if database_file is not None:
-                self.database_name = database_file
-            else:
-                parameters_file = tempfile.NamedTemporaryFile(mode="w", delete=False, dir=None, suffix=".sqlite")
-                parameters_file.close()
-                self.database_name = parameters_file.name
+        # remove database and create structure
+        if remove_existing and os.path.exists(self.database_name):
+            os.remove(self.database_name)
+            self.create_structure()
+
+        # set datastore to problem
+        self.problem.data_store = self
+
+        # set db log
+        if self.problem.options['log_db_handler']:
+            # create file handler and set level to debug
+            file_handler = SqliteHandler(self)
+            file_handler.setLevel(logging.DEBUG)
+            # add formatter to SqliteHandler
+            file_handler.setFormatter(self.problem.formatter)
+            # add SqliteHandler to logger
+            self.problem.logger.addHandler(file_handler)
 
     def __del__(self):
         super().__del__()
@@ -184,12 +189,11 @@ class SqliteDataStore(DataStore):
                    'initial_value NUMBER, low_boundary NUMBER, high_boundary NUMBER, precision NUMBER);'
         self._execute_command(exec_cmd)
 
-        for parameter in self.problem.get_parameters_list():
-            parameter_arg = [0] * 5
-            length = len(parameter)
-            parameter_arg[0:length] = parameter
-            exec_cmd = 'INSERT INTO vector(name, initial_value, low_boundary, high_boundary, precision) ' \
-                       "VALUES( '%s', %f, %f, %f, %f);" % tuple(parameter_arg)
+        for parameter in self.problem.parameters.items():
+            exec_cmd = "INSERT INTO vector(name, initial_value, low_boundary, high_boundary, precision) " \
+                       "VALUES( '{}', {}, {}, {}, {});".format(parameter[0], parameter[1]['initial_value'],
+                                                               parameter[1]['bounds'][0], parameter[1]['bounds'][1],
+                                                               parameter[1]['precision'] if 'precision' in parameter[1] else 0)
             self._execute_command(exec_cmd)
 
     def _create_structure_costs(self):
@@ -270,7 +274,7 @@ class SqliteDataStore(DataStore):
         connection.close()
         """
 
-    def read_problem(self, problem):
+    def read_from_datastore(self):
         connection = sqlite3.connect(self.database_name)
         # connection.execute('pragma journal_mode=wal')
 
@@ -278,25 +282,25 @@ class SqliteDataStore(DataStore):
 
         exec_cmd_problem = "SELECT * FROM problem"
         problem_table = cursor.execute(exec_cmd_problem).fetchall()
-        problem.name = problem_table[0][1]
+        self.problem.name = problem_table[0][1]
 
-        # vector
+        # parameters
         exec_cmd_params = "SELECT * FROM parameters"
-
         columns = []
         for row in cursor.execute(exec_cmd_params):
             columns.append(row)
 
-        problem.parameters = {}
+        self.problem.parameters = {}
         for parameter in columns:
-            problem.parameters[parameter[1]] = {'initial_value': parameter[2],
+            self.problem.parameters[parameter[1]] = {'initial_value': parameter[2],
                                                 'bounds': [parameter[3], parameter[4]], 'precision': parameter[5]}
         exec_cmd_costs = "SELECT * FROM costs"
 
+        # costs
         costs = []
         for cost in cursor.execute(exec_cmd_costs):
             costs.append(cost[1])
-        problem.costs = {cost: 0.0 for cost in costs}
+        self.problem.costs = {cost: 0.0 for cost in costs}
 
         exec_cmd_data = "SELECT * FROM data"
         data = cursor.execute(exec_cmd_data)
@@ -320,9 +324,9 @@ class SqliteDataStore(DataStore):
         current_population = table[0][1]
         for row in table:
             if row[1] == current_population:
-                individual = Individual(row[2:2 + len(problem.parameters)])
-                l = 2 + len(problem.parameters) + len(problem.costs)
-                individual.costs = row[2 + len(problem.parameters): 2 + len(problem.parameters) + len(problem.costs)]
+                individual = Individual(row[2:2 + len(self.problem.parameters)])
+                l = 2 + len(self.problem.parameters) + len(self.problem.costs)
+                individual.costs = row[2 + len(self.problem.parameters): 2 + len(self.problem.parameters) + len(self.problem.costs)]
                 individual.front_number = row[l]
                 individual.crowding_distance = row[l+1]
                 individual.feasible = row[l+2]
@@ -344,6 +348,7 @@ class DummyDataStore(DataStore):
 
     def __init__(self, problem=None):
         super().__init__(problem)
+        self.create_structure()
 
     def __del__(self):
         super().__del__()
